+++ Capitalism/com/indong/capitalism/Frame/CustomTable/CMainPanelTableModel.java
@@ -10,7 +10,7 @@ public class CMainPanelTableModel extends DefaultTableModel{
         return 1;
     }
 
-    public void addRow(String name,int allAsset , int deposit , int cash , int loan) {
+    public void addRow(String name,long allAsset , long deposit , long cash , long loan) {
         super.addRow(new Object[]{new CMainTableComp(name,allAsset , deposit , cash , loan)});//이름 총자산 예금 현금 대출
     }
 }
